requirements.txt
# Core LangChain dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-experimental==0.0.51

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Visualization (optional, for advanced analysis)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Environment management
python-dotenv>=1.0.0

